ccmlib/cmds/node_cmds: make versionfrombuild useful
make `versionfrombuild` command work without a running node
we don't need to cross check with nodetool that the version is
the correct one, the once mention in scylla package is enough
for us.

we need this working for driver test that would be able
to pick test base on the versions.